refactor: 파일 이름 변경 (#51)
- login 관련 스키마 파일들을 auth 디렉토리에서 login 디렉토리로 이름 변경
