+++ src/features/auth/api/schemas/login/password-reset.ts
@@ -1,5 +1,5 @@
 import z from 'zod'
-import { ErrorResponseSchema } from '@/features/auth/api/schemas/auth/common'
+import { ErrorResponseSchema } from '@/features/auth/api/schemas/login/common'
 
 export const PasswordResetRequestSchema = z.object({
   reset_token: z.string().min(1),
